requirements.txt
openai
sounddevice
soundfile
numpy
python-dotenv
thefuzz
python-Levenshtein
pandas
